Revert "HIVE-25967: Prevent residual expressions from getting serialized in Iceberg splits (#3041) (Adam Szita, reviewed by Peter Vary)"
This reverts commit 1aa6ce800004798e78ea53c3bec2beedb5f55b6c.
